# bigbike-backend/src/main/resources/openapi/bigbike-openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "BigBike Backend API",
    "version": "1.0.0",
    "description": "REST API for BigBike — Vietnam's premium motorcycle marketplace. Rewrite of WordPress backend. All endpoints return ApiDataResponse<T> or ApiListResponse<T> envelopes."
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development"
    },
    {
      "url": "https://api.bigbike.vn",
      "description": "Production"
    }
  ],
...
  "paths": {
...
    "/api/v1/admin/customers": {
      "get": {
        "tags": [
          "Admin Customers"
        ],
        "summary": "List customers (admin)",
        "operationId": "adminListCustomers",
        "security": [
          {
            "AdminBearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "size",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20
            }
          },
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Customer list"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/v1/admin/customers/{customerId}": {
      "get": {
        "tags": [
          "Admin Customers"
        ],
        "summary": "Get customer by id (admin)",
        "operationId": "adminGetCustomer",
        "security": [
          {
            "AdminBearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "customerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Customer detail"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      },
      "patch": {
        "tags": [
          "Admin Customers"
        ],
        "summary": "Update customer (admin)",
        "operationId": "adminUpdateCustomer",
        "security": [
          {
            "AdminBearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/v1/admin/customers/{customerId}/status": {
      "patch": {
        "tags": [
          "Admin Customers"
        ],
        "summary": "Update customer status (admin)",
        "operationId": "adminUpdateCustomerStatus",
        "security": [
          {
            "AdminBearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
    "/api/v1/admin/reports/analytics": {
      "get": {
        "tags": [
          "Admin Reports"
        ],
        "summary": "Get analytics report (admin)",
        "operationId": "adminReportsAnalytics",
        "security": [
          {
            "AdminBearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Analytics report",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminAnalyticsResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/v1/admin/reports/orders/export": {
      "get": {
        "tags": [
          "Admin Reports"
        ],
        "summary": "Export orders report CSV (admin)",
        "operationId": "adminReportsOrdersExport",
        "security": [
          {
            "AdminBearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "CSV export"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/v1/admin/reports/customers/export": {
      "get": {
        "tags": [
          "Admin Reports"
        ],
        "summary": "Export customers report CSV (admin)",
        "operationId": "adminReportsCustomersExport",
        "security": [
          {
            "AdminBearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "CSV export"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/v1/admin/reports/products/export": {
      "get": {
        "tags": [
          "Admin Reports"
        ],
        "summary": "Export products report CSV (admin)",
        "operationId": "adminReportsProductsExport",
        "security": [
          {
            "AdminBearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "publishStatus",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "CSV export"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "AdminBearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Admin JWT access token. Obtain via POST /api/v1/auth/login. Header: Authorization: Bearer <accessToken>"
      },
      "CustomerSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "bb_session",
        "description": "Customer session cookie. Set automatically after POST /api/v1/customer/auth/login."
      },
      "CsrfHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token",
        "description": "CSRF token required for all cart/checkout mutations. Value must match the bb_csrf cookie set by the server."
      }
    },
    "responses": {
...
      "Unauthorized": {
        "description": "Authentication required or token expired",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiErrorResponse"
            }
          }
        }
      },
...
    },
    "schemas": {
...
      "AdminAnalyticsResponse": {
        "type": "object",
        "properties": {
          "summary": {
            "$ref": "#/components/schemas/AdminAnalyticsPeriodSummary"
          },
          "dailyRevenue": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AdminAnalyticsDailyRevenue"
            }
          },
          "topProducts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AdminAnalyticsTopProduct"
            }
          },
          "topCustomers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AdminAnalyticsTopCustomer"
            }
          }
        }
      },
      "AdminAnalyticsPeriodSummary": {
        "type": "object",
        "description": "Aggregate revenue metrics for the requested date range. All amounts in VND (BigDecimal serialized with scale 2).",
        "properties": {
          "grossOrderValue": {
            "type": "number",
            "description": "GMV: SUM(totalAmount) excluding CANCELLED orders."
          },
          "paidRevenue": {
            "type": "number",
            "description": "Recognized revenue: SUM(totalAmount) where status = COMPLETED. Response name is retained for compatibility."
          },
          "orderCount": {
            "type": "integer",
            "description": "COUNT of orders excluding CANCELLED."
          },
          "avgOrderValue": {
            "type": "number",
            "description": "grossOrderValue / orderCount. Zero when orderCount is 0."
          }
        }
      },
      "AdminAnalyticsDailyRevenue": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "format": "date",
            "description": "ISO-8601 date YYYY-MM-DD in Asia/Ho_Chi_Minh timezone."
          },
          "revenue": {
            "type": "number",
            "description": "Daily grossOrderValue (same REVENUE_EXCLUDED set as summary)."
          }
        }
      },
      "AdminAnalyticsTopProduct": {
        "type": "object",
        "properties": {
          "productKey": {
            "type": "string",
            "description": "COALESCE(product_pk, product_id::text). Stable identifier covering both admin-created and regular products."
          },
          "productName": {
            "type": "string",
            "description": "Product name snapshot from order line item."
          },
          "revenue": {
            "type": "number",
            "description": "SUM(lineTotal) excluding CANCELLED orders."
          },
          "unitsSold": {
            "type": "integer",
            "description": "SUM(quantity) excluding RANKING_EXCLUDED order statuses."
          }
        }
      },
      "AdminAnalyticsTopCustomer": {
        "type": "object",
        "properties": {
          "customerKey": {
            "type": "string",
            "description": "COALESCE(customer_id::text, customer_email). Stable group key preventing split rows on email change."
          },
          "customerEmail": {
            "type": "string",
            "description": "MAX(customer_email) — display email."
          },
          "revenue": {
            "type": "number",
            "description": "SUM(totalAmount) excluding RANKING_EXCLUDED order statuses."
          },
          "orderCount": {
            "type": "integer",
            "description": "COUNT of orders excluding RANKING_EXCLUDED order statuses."
          }
        }
      },
      "ApiMeta": {
        "type": "object",
        "properties": {
          "requestId": {
            "type": "string",
            "format": "uuid"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "ApiErrorResponse": {
        "type": "object",
        "description": "Standard error envelope used by all error responses",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiError"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          }
        }
      },
      "ApiError": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "VALIDATION_ERROR"
          },
          "message": {
            "type": "string",
            "example": "Validation failed."
          },
          "details": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ApiErrorDetail"
            }
          }
        }
      },
      "ApiErrorDetail": {
        "type": "object",
        "properties": {
          "field": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      }
    }
  }
}